override ActiveAdmin::Namespace's register_module method to avoid Admin Namespace confliction with Admin model
By default, the undefined namespace modules of ActiveAdmin are defined by Object.const_set in ActiveAdmin::Namespace#register_module, this will cause Rails autoloading to lose its functionality and breaks the definition of a object having the same name. The workaround simply skips the registration of 'Admin' model and let Rails to autoload it.